# docs/openapi/climate-api-v1.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "Climate Monitoring System API",
                 "description":  "Public API exposed by Climate.Gateway. Internal service endpoints are intentionally excluded.",
                 "version":  "v1"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Local Docker Gateway"
                    }
                ],
    "paths":  {
...
                  "/api/users/me":  {
                                        "get":  {
                                                    "tags":  [
                                                                 "Users"
                                                             ],
                                                    "responses":  {
                                                                      "200":  {
                                                                                  "description":  "OK",
                                                                                  "content":  {
                                                                                                  "text/plain":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/UserResponse"
                                                                                                                                }
                                                                                                                 },
                                                                                                  "application/json":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/UserResponse"
                                                                                                                                      }
                                                                                                                       },
                                                                                                  "text/json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/UserResponse"
                                                                                                                               }
                                                                                                                }
                                                                                              }
                                                                              },
                                                                      "401":  {
                                                                                  "description":  "Unauthorized",
                                                                                  "content":  {
                                                                                                  "text/plain":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                }
                                                                                                                 },
                                                                                                  "application/json":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                      }
                                                                                                                       },
                                                                                                  "text/json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                               }
                                                                                                                }
                                                                                              }
                                                                              },
                                                                      "403":  {
                                                                                  "description":  "Forbidden"
                                                                              }
                                                                  },
                                                    "operationId":  "Users_get_api_users_me",
                                                    "security":  [
                                                                     {
                                                                         "Bearer":  [

                                                                                    ]
                                                                     }
                                                                 ]
                                                }
                                    },
                  "/api/users":  {
                                     "get":  {
                                                 "tags":  [
                                                              "Users"
                                                          ],
                                                 "responses":  {
                                                                   "200":  {
                                                                               "description":  "OK",
                                                                               "content":  {
                                                                                               "text/plain":  {
                                                                                                                  "schema":  {
                                                                                                                                 "type":  "array",
                                                                                                                                 "items":  {
                                                                                                                                               "$ref":  "#/components/schemas/UserResponse"
                                                                                                                                           }
                                                                                                                             }
                                                                                                              },
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "type":  "array",
                                                                                                                                       "items":  {
                                                                                                                                                     "$ref":  "#/components/schemas/UserResponse"
                                                                                                                                                 }
                                                                                                                                   }
                                                                                                                    },
                                                                                               "text/json":  {
                                                                                                                 "schema":  {
                                                                                                                                "type":  "array",
                                                                                                                                "items":  {
                                                                                                                                              "$ref":  "#/components/schemas/UserResponse"
                                                                                                                                          }
                                                                                                                            }
                                                                                                             }
                                                                                           }
                                                                           },
                                                                   "401":  {
                                                                               "description":  "Unauthorized"
                                                                           },
                                                                   "403":  {
                                                                               "description":  "Forbidden"
                                                                           }
                                                               },
                                                 "operationId":  "Users_get_api_users",
                                                 "security":  [
                                                                  {
                                                                      "Bearer":  [

                                                                                 ]
                                                                  }
                                                              ]
                                             }
                                 },
...
                  "/api/users/{id}/status":  {
                                                 "patch":  {
                                                               "tags":  [
                                                                            "Users"
                                                                        ],
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "id",
                                                                                      "in":  "path",
                                                                                      "required":  true,
                                                                                      "schema":  {
                                                                                                     "type":  "string",
                                                                                                     "format":  "uuid"
                                                                                                 }
                                                                                  }
                                                                              ],
                                                               "requestBody":  {
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/UpdateUserStatusRequest"
                                                                                                                                       }
                                                                                                                        },
                                                                                                   "text/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/UpdateUserStatusRequest"
                                                                                                                                }
                                                                                                                 },
                                                                                                   "application/*+json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/UpdateUserStatusRequest"
                                                                                                                                         }
                                                                                                                          }
                                                                                               }
                                                                               },
                                                               "responses":  {
                                                                                 "204":  {
                                                                                             "description":  "No Content"
                                                                                         },
                                                                                 "401":  {
                                                                                             "description":  "Unauthorized"
                                                                                         },
                                                                                 "403":  {
                                                                                             "description":  "Forbidden"
                                                                                         }
                                                                             },
                                                               "operationId":  "Users_patch_api_users_id_status",
                                                               "security":  [
                                                                                {
                                                                                    "Bearer":  [

                                                                                               ]
                                                                                }
                                                                            ]
                                                           }
                                             },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "ProblemDetails":  {
                                                              "type":  "object",
                                                              "properties":  {
                                                                                 "type":  {
                                                                                              "type":  "string",
                                                                                              "nullable":  true
                                                                                          },
                                                                                 "title":  {
                                                                                               "type":  "string",
                                                                                               "nullable":  true
                                                                                           },
                                                                                 "status":  {
                                                                                                "type":  "integer",
                                                                                                "format":  "int32",
                                                                                                "nullable":  true
                                                                                            },
                                                                                 "detail":  {
                                                                                                "type":  "string",
                                                                                                "nullable":  true
                                                                                            },
                                                                                 "instance":  {
                                                                                                  "type":  "string",
                                                                                                  "nullable":  true
                                                                                              }
                                                                             },
                                                              "additionalProperties":  {

                                                                                       }
                                                          },
...
                                       "UpdateUserStatusRequest":  {
                                                                       "type":  "object",
                                                                       "properties":  {
                                                                                          "isActive":  {
                                                                                                           "type":  "boolean"
                                                                                                       }
                                                                                      },
                                                                       "additionalProperties":  false
                                                                   },
                                       "UserResponse":  {
                                                            "type":  "object",
                                                            "properties":  {
                                                                               "id":  {
                                                                                          "type":  "string",
                                                                                          "format":  "uuid"
                                                                                      },
                                                                               "username":  {
                                                                                                "type":  "string",
                                                                                                "nullable":  true
                                                                                            },
                                                                               "email":  {
                                                                                             "type":  "string",
                                                                                             "nullable":  true
                                                                                         },
                                                                               "role":  {
                                                                                            "type":  "string",
                                                                                            "nullable":  true
                                                                                        },
                                                                               "isActive":  {
                                                                                                "type":  "boolean"
                                                                                            },
                                                                               "createdAt":  {
                                                                                                 "type":  "string",
                                                                                                 "format":  "date-time"
                                                                                             },
                                                                               "updatedAt":  {
                                                                                                 "type":  "string",
                                                                                                 "format":  "date-time"
                                                                                             }
                                                                           },
                                                            "additionalProperties":  false
                                                        },
...
                                   },
                       "securitySchemes":  {
                                               "Bearer":  {
                                                              "type":  "http",
                                                              "scheme":  "bearer",
                                                              "bearerFormat":  "JWT",
                                                              "description":  "Paste only the accessToken returned by POST /api/auth/login. Do not include the 'Bearer ' prefix; Swagger UI adds it automatically."
                                                          }
                                           }
                   },
...
}